requirements.txt
torch>=2.0.0
transformers>=4.38.0
datasets>=2.15.0
sentence-transformers>=2.2.0
huggingface-hub>=0.19.0
wandb>=0.16.0
python-dotenv>=1.0.0
tqdm>=4.66.0
flash-attn>=2.5.0; platform_system == "Linux"
